# Conflicts:
#	features/merchant/seller_tokopedia_flash_sale/src/main/java/com/tokopedia/tkpd/flashsale/presentation/detail/CampaignDetailFragment.kt

+++ features/content/feed_component/src/main/java/com/tokopedia/feedcomponent/view/viewmodel/track/TrackingViewModel.kt
@@ -9,7 +9,7 @@ class TrackingViewModel(
         val type: String = "",
         val source: String = "",
         val viewType: String = "",
-        val recomId: Int = 0
+        val recomId: Long = 0
 ) {
     fun copy(): TrackingViewModel {
         return TrackingViewModel(clickURL, viewURL, type, source)
